fix(graphs): preserve permission state on theme changes
