+++ src/components/router/visualization.tsx
@@ -37,6 +37,7 @@ import {
 import { RouterRoutesResponse } from '@/lib/models/Router';
 import { useChartColors } from '@/lib/hooks/useChartColors';
 import { getReactFlowMarkerColor } from '@/lib/semantic-colors';
+import { mergeSemanticEdgeColors } from '@/lib/reactflow-edge-colors';
 
 const getHttpMethodColor = (method: string) => {
   switch (method.toUpperCase()) {
@@ -57,45 +58,6 @@ const getHttpMethodColor = (method: string) => {
     default:
       return 'bg-surface-3 text-foreground-muted border-border-strong hover:bg-surface-3';
   }
-};
-
-const mergeSemanticEdgeColors = (
-  currentEdge: Edge,
-  semanticEdge: Edge
-): Edge => {
-  let updatedEdge = currentEdge;
-
-  if (semanticEdge.style?.stroke !== undefined) {
-    updatedEdge = {
-      ...updatedEdge,
-      style: {
-        ...currentEdge.style,
-        stroke: semanticEdge.style.stroke,
-      },
-    };
-  }
-
-  if (semanticEdge.labelStyle?.fill !== undefined) {
-    updatedEdge = {
-      ...updatedEdge,
-      labelStyle: {
-        ...currentEdge.labelStyle,
-        fill: semanticEdge.labelStyle.fill,
-      },
-    };
-  }
-
-  if (currentEdge.markerEnd && semanticEdge.markerEnd?.color !== undefined) {
-    updatedEdge = {
-      ...updatedEdge,
-      markerEnd: {
-        ...currentEdge.markerEnd,
-        color: semanticEdge.markerEnd.color,
-      },
-    };
-  }
-
-  return updatedEdge;
 };
 
 // Memoized custom node components for better performance
@@ -655,17 +617,8 @@ export const RouterVisualization = ({ data }: Props) => {
   const [edges, setEdges, onEdgesChange] = useEdgesState(initialEdges);
 
   useEffect(() => {
-    const semanticEdgesById = new Map(
-      initialEdges.map(edge => [edge.id, edge] as const)
-    );
-
     setEdges(currentEdges =>
-      currentEdges.map(currentEdge => {
-        const semanticEdge = semanticEdgesById.get(currentEdge.id);
-        return semanticEdge
-          ? mergeSemanticEdgeColors(currentEdge, semanticEdge)
-          : currentEdge;
-      })
+      mergeSemanticEdgeColors(currentEdges, initialEdges)
     );
   }, [initialEdges, setEdges]);
